Currency convertor

+++ Currency Converter/task/cconverter/cconverter.py
@@ -1,3 +1,3 @@
 coni = int(input('Please, enter the number of conicoins you have: '))
-rate = int(input('Please, enter the exchange rate: '))
+rate = float(input('Please, enter the exchange rate: '))
 print(f'The total amount of dollars: {coni * rate}')
